+++ samples/showcase/music.preview.dali.cpp
@@ -17,7 +17,7 @@ View e6=View::New(); e6.SetBackgroundColor(UiColor(0xA855F7)); e6.SetCornerRadiu
 View e7=View::New(); e7.SetBackgroundColor(UiColor(0x818CF8)); e7.SetCornerRadius(7.0f); e7.SetRequestedWidth(26.0f); e7.SetRequestedHeight(190.0f); e7.SetMargin(Extents(7,7,0,0));
 View e8=View::New(); e8.SetBackgroundColor(UiColor(0x6366F1)); e8.SetCornerRadius(7.0f); e8.SetRequestedWidth(26.0f); e8.SetRequestedHeight(95.0f); e8.SetMargin(Extents(7,7,0,0));
 View e9=View::New(); e9.SetBackgroundColor(UiColor(0xA855F7)); e9.SetCornerRadius(7.0f); e9.SetRequestedWidth(26.0f); e9.SetRequestedHeight(155.0f); e9.SetMargin(Extents(7,7,0,0));
-eq.AddChildren({ e1,e2,e3,e4,e5,e6,e7,e8,e9 });
+eq.Add(e1); eq.Add(e2); eq.Add(e3); eq.Add(e4); eq.Add(e5); eq.Add(e6); eq.Add(e7); eq.Add(e8); eq.Add(e9);
 
 Label now = Label::New("NOW PLAYING"); now.SetFontSize(24); now.SetTextColor(UiColor(0x818CF8)); now.SetMargin(Extents(0,0,0,12));
 Label song = Label::New("Midnight City"); song.SetFontSize(72); song.SetTextColor(UiColor(0xF8FAFC));
@@ -27,16 +27,16 @@ View bar = View::New(); bar.SetBackgroundColor(UiColor(0x27273B)); bar.SetCorner
 bar.SetRequestedWidth(760.0f); bar.SetRequestedHeight(12.0f); bar.SetMargin(Extents(0,0,36,0));
 View prog = View::New(); prog.SetBackgroundColor(UiColor(0xEC4899)); prog.SetCornerRadius(6.0f);
 prog.SetRequestedWidth(470.0f); prog.SetRequestedHeight(12.0f);
-bar.AddChildren({ prog });
+bar.Add(prog);
 
 FlexLayout ctr = FlexLayout::New(); ctr.SetDirection(FlexDirection::ROW); ctr.SetAlignItems(FlexAlign::CENTER);
 View pv = View::New(); pv.SetBackgroundColor(UiColor(0x27273B)); pv.SetCornerRadius(36.0f); pv.SetRequestedWidth(72.0f); pv.SetRequestedHeight(72.0f); pv.SetMargin(Extents(0,28,0,0));
 FlexLayout play = FlexLayout::New(); play.SetJustifyContent(FlexJustify::CENTER); play.SetAlignItems(FlexAlign::CENTER);
 play.SetBackgroundColor(UiColor(0xEC4899)); play.SetCornerRadius(56.0f); play.SetRequestedWidth(112.0f); play.SetRequestedHeight(112.0f); play.SetMargin(Extents(28,28,0,0));
 Label pl = Label::New("II"); pl.SetFontSize(44); pl.SetTextColor(UiColor(0xFFFFFF));
-play.AddChildren({ pl });
+play.Add(pl);
 View nx = View::New(); nx.SetBackgroundColor(UiColor(0x27273B)); nx.SetCornerRadius(36.0f); nx.SetRequestedWidth(72.0f); nx.SetRequestedHeight(72.0f); nx.SetMargin(Extents(28,0,0,0));
-ctr.AddChildren({ pv, play, nx });
+ctr.Add(pv); ctr.Add(play); ctr.Add(nx);
 
-root.AddChildren({ eq, now, song, artist, bar, ctr });
+root.Add(eq); root.Add(now); root.Add(song); root.Add(artist); root.Add(bar); root.Add(ctr);
 return root;